# docs/generated/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orion API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/approval-requests/{id}": {
      "get": {
        "operationId": "getApprovalRequest",
        "tags": ["Approval Requests"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "creatorId",
                    "title",
                    "description",
                    "status",
                    "version",
                    "rejectionReason",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
                    },
                    "creatorId": {
                      "type": "string",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200,
                      "pattern": ".*\\S.*"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000,
                          "pattern": ".*\\S.*"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "const": "DRAFT"
                        },
                        {
                          "type": "string",
                          "const": "SUBMITTED"
                        },
                        {
                          "type": "string",
                          "const": "APPROVED"
                        },
                        {
                          "type": "string",
                          "const": "REJECTED"
                        },
                        {
                          "type": "string",
                          "const": "CANCELLED"
                        }
                      ]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "rejectionReason": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000,
                          "pattern": ".*\\S.*"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "x-expected-error-codes": [
          "VALIDATION_FAILED",
          "AUTHENTICATION_REQUIRED",
          "RESOURCE_NOT_FOUND"
        ]
      }
    },
...
    "/approval-requests/{id}/submit": {
      "post": {
        "operationId": "submitApprovalRequest",
        "tags": ["Approval Requests"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["expectedVersion"],
                "properties": {
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "creatorId",
                    "title",
                    "description",
                    "status",
                    "version",
                    "rejectionReason",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
                    },
                    "creatorId": {
                      "type": "string",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200,
                      "pattern": ".*\\S.*"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000,
                          "pattern": ".*\\S.*"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "const": "DRAFT"
                        },
                        {
                          "type": "string",
                          "const": "SUBMITTED"
                        },
                        {
                          "type": "string",
                          "const": "APPROVED"
                        },
                        {
                          "type": "string",
                          "const": "REJECTED"
                        },
                        {
                          "type": "string",
                          "const": "CANCELLED"
                        }
                      ]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "rejectionReason": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000,
                          "pattern": ".*\\S.*"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "requestId"],
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Z][A-Z0-9_]*$"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "traceId": {
                          "type": "string"
                        },
                        "errorId": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "x-expected-error-codes": [
          "VALIDATION_FAILED",
          "AUTHENTICATION_REQUIRED",
          "RESOURCE_NOT_FOUND",
          "APPROVAL_REQUEST_INVALID_STATE",
          "RESOURCE_VERSION_CONFLICT"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}